Further handling of user input. Dealing with negative numbers and decimal points

+++ app/src/main/java/elr/calculatortest1/io/EightInputHandlerImpl.java
@@ -9,7 +9,7 @@ public class EightInputHandlerImpl extends NumberInputHandler {
     }
 
     @Override
-    protected double getInputValue() {
-        return 8.0;
+    protected int getInputValue() {
+        return 8;
     }
 }
